net/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
mysql-connector-python==9.0.0
joblib==1.4.2
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
scapy==2.5.0
netmiko==4.4.0
paramiko==3.5.0
textfsm==1.1.3
